requirements.txt
timm>=0.9.2
torch>=2.0.0
torchvision>=0.15.0
matplotlib>=3.7.0
numpy>=1.24.0
Pillow>=9.5.0
yt-dlp>=2023.12.30 